ARUHA-282: add toogle to switch off creation and deletion of event type

+++ src/main/java/de/zalando/aruha/nakadi/controller/EventTypeController.java
@@ -16,9 +16,7 @@ import de.zalando.aruha.nakadi.partitioning.PartitionResolver;
 import de.zalando.aruha.nakadi.problem.ValidationProblem;
 import de.zalando.aruha.nakadi.repository.EventTypeRepository;
 import de.zalando.aruha.nakadi.repository.TopicRepository;
-import java.util.List;
-import java.util.Objects;
-import javax.validation.Valid;
+import de.zalando.aruha.nakadi.util.FeatureToggleService;
 import org.everit.json.schema.SchemaException;
 import org.everit.json.schema.loader.SchemaLoader;
 import org.json.JSONException;
@@ -28,7 +26,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import static org.springframework.http.ResponseEntity.status;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -36,6 +33,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.NativeWebRequest;
+
+import javax.validation.Valid;
+import java.util.List;
+import java.util.Objects;
+
+import static org.springframework.http.ResponseEntity.status;
 import static org.zalando.problem.spring.web.advice.Responses.create;
 
 @RestController
@@ -48,16 +51,20 @@ public class EventTypeController {
     private final TopicRepository topicRepository;
     private final PartitionResolver partitionResolver;
     private final Enrichment enrichment;
+    private final FeatureToggleService featureToggleService;
 
     @Autowired
     public EventTypeController(final EventTypeRepository eventTypeRepository,
                                final TopicRepository topicRepository,
                                final PartitionResolver partitionResolver,
-                               final Enrichment enrichment) {
+                               final Enrichment enrichment,
+                               final FeatureToggleService featureToggleService)
+    {
         this.eventTypeRepository = eventTypeRepository;
         this.topicRepository = topicRepository;
         this.partitionResolver = partitionResolver;
         this.enrichment = enrichment;
+        this.featureToggleService = featureToggleService;
     }
 
     @RequestMapping(method = RequestMethod.GET)
@@ -70,7 +77,11 @@ public class EventTypeController {
     @RequestMapping(method = RequestMethod.POST)
     public ResponseEntity<?> createEventType(@Valid @RequestBody final EventType eventType,
                                              final Errors errors,
-                                             final NativeWebRequest nativeWebRequest) {
+                                             final NativeWebRequest nativeWebRequest)
+    {
+        if (featureToggleService.isFeatureEnabled("disable_event_type_creation")) {
+            return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+        }
         if (errors.hasErrors()) {
             return create(new ValidationProblem(errors), nativeWebRequest);
         }
@@ -103,7 +114,11 @@ public class EventTypeController {
 
     @RequestMapping(value = "/{name:.+}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteEventType(@PathVariable("name") final String eventTypeName,
-                                             final NativeWebRequest nativeWebRequest) {
+                                             final NativeWebRequest nativeWebRequest)
+    {
+        if (featureToggleService.isFeatureEnabled("disable_event_type_deletion")) {
+            return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
+        }
         try {
             eventTypeRepository.removeEventType(eventTypeName);
             topicRepository.deleteTopic(eventTypeName);
